% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_sampler_DLM.R
\name{Gibbs_sampler}
\alias{Gibbs_sampler}
\title{Fit DLM with MCMC}
\usage{
Gibbs_sampler(x, y, z, df_l, niter, burnin, tau2, xi2, a, b)
}
\arguments{
\item{x}{matrix of exposure history (columns) for individuals (rows) (class "\code{matrix}")}

\item{y}{vector of response values (class "\code{numeric}")}

\item{z}{matrix of covariates, not including the modifiers (class "\code{matrix}")}

\item{df_l}{degrees of freedom for exposure time basis (class "\code{numeric}")}

\item{niter}{number of MCMC iterations class "\code{numeric}")}

\item{burnin}{number of MCMC iterations to remove as warm-up class "\code{numeric}")}

\item{tau2}{prior variance for cross-basis regression coefficients (class "\code{numeric}")}

\item{xi2}{prior variance for covariate and modifier regression coefficients (class "\code{numeric}")}

\item{a}{prior shape for error variance for Gaussian model class "\code{numeric}")}

\item{b}{prior scale for error variance for Gaussian model class "\code{numeric}")}
}
\value{
This function returns a list that is an object of class "\code{dlimIM}" with the following components
\item{posterior}{posterior samples including burn-in (class "\code{matrix}")}
}
\description{
Fit DLM with MCMC
}
