% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_sampler_m.R
\name{MCMC_sampler_m}
\alias{MCMC_sampler_m}
\title{Fit DLIM-IM}
\usage{
MCMC_sampler_m(
  x,
  y,
  trials = NULL,
  M,
  z,
  df_m,
  df_l,
  tau2,
  xi2,
  a = NULL,
  b = NULL,
  niter,
  burnin,
  model_type = "ns",
  var_select = FALSE,
  weights_prior = NULL,
  WAIC = FALSE,
  prior_inclusions = NULL,
  family = "gaussian",
  nchains = 1
)
}
\arguments{
\item{x}{matrix of exposure history (columns) for individuals (rows) (class "\code{matrix}")}

\item{y}{vector of response values (class "\code{numeric}")}

\item{trials}{number of trials for each observation when \code{family = "binomial"} (class "\code{numeric}")}

\item{M}{numeric matrix of modifying values (columns) for individuals with values between 0 and 1 (rows) (class "\code{matrix}")}

\item{z}{matrix of covariates, not including the modifiers (class "\code{matrix}")}

\item{df_m}{degrees of freedom for modifier basis (class "\code{numeric}")}

\item{df_l}{degrees of freedom for exposure time basis (class "\code{numeric}")}

\item{tau2}{prior variance for cross-basis regression coefficients (class "\code{numeric}")}

\item{xi2}{prior variance for covariate and modifier regression coefficients (class "\code{numeric}")}

\item{a}{prior shape for error variance for Gaussian model class "\code{numeric}")}

\item{b}{prior scale for error variance for Gaussian model class "\code{numeric}")}

\item{niter}{number of MCMC iterations class "\code{numeric}")}

\item{burnin}{number of MCMC iterations to remove as warm-up class "\code{numeric}")}

\item{model_type}{modifier spline basis construction ("linear" or "ns" for natural splines) class "\code{character}")}

\item{var_select}{TRUE to indicate performing selection on the modifiers class "\code{logical}")}

\item{weights_prior}{vector of Dirichlet shape parameters for each modifier for their prior distribution class "\code{numeric}")}

\item{WAIC}{TRUE to indicate computing WAIC class "\code{logical}")}

\item{prior_inclusions}{vector of prior inclusion probabilities for each weight class "\code{numeric}")}

\item{family}{family object specifying the distribution in fitting, either 'gaussian' or 'binomial' class "\code{character}")}

\item{nchains}{number of chains}
}
\value{
This function returns a list that is an object of class "\code{dlimIM}" with the following components
\item{posterior}{posterior samples including burn-in (class "\code{matrix}")}
}
\description{
Create posterior samples from a DLIM-IM
}
\seealso{
Type \code{vignette('dlimOverview')} for a detailed description.

\link[dlimIM]{pred_m}

\link[dlimIM]{plot_bdlim}
}
