% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bdlim.R
\name{plot_bdlim}
\alias{plot_bdlim}
\title{Plotting for DLIM-IM}
\usage{
plot_bdlim(
  x,
  pred = NULL,
  m_star,
  sel,
  type,
  exp_times = NULL,
  time_pts = NULL,
  n_col = 3
)
}
\arguments{
\item{x}{posterior samples  (class "\code{dlimIM}")}

\item{pred}{prediction object from \link[dlimIM]{pred_m}}

\item{m_star}{vector of modifier index values (class "\code{numeric}")}

\item{sel}{selects which iterations of the MCMC sampler to use for inference (class "\code{numeric}")}

\item{type}{plot type options: "cumulative", "by_time", "by_modifier" (class "\code{character}")}

\item{exp_times}{labels for exposure time points (class "\code{character}")}

\item{time_pts}{exposure time points to plot by when \code{plot_by = "by_time"} (class "\code{numeric}")}

\item{n_col}{number of columns for plotting grid (class "\code{numeric}")}
}
\value{
This function returns a ggplot of specified \code{type}
}
\description{
Create cumulative and point-wise effect plots for a DLIM-IM
}
\seealso{
\link[dlimIM]{pred_m}

\link[dlimIM]{MCMC_sampler_m}
}
