% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_01.R
\name{scale_01}
\alias{scale_01}
\title{scale to (0,1) interval}
\usage{
scale_01(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
This function returns the vector \code{x} scaled to lie within the (0,1) interval
}
\description{
scale to (0,1) interval
}
