% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlimIM.R
\docType{package}
\name{dlimIM}
\alias{dlimIM}
\title{dlimIM Package}
\description{
Collection of functions for fitting and interpreting distributed lag interaction models with index modification (DLIM-IM). A DLIM-IM regresses a scalar outcome on repeated measures of exposure and allows for modification by an index constructed from multiple candidate modifiers. Includes a MCMC_sampler_m() function for sampling from the posterior, pred_m() function for inference, and plotting functions for visualization.
}
\author{
Danielle Demateis

Ander Wilson

Kayleigh Keller
}
