% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_m.R
\name{pred_m}
\alias{pred_m}
\title{Prediction for DLIM-IM}
\usage{
pred_m(
  posterior_list,
  chain = 1,
  burnin,
  thin = 1,
  new_mods = NULL,
  m_star = NULL,
  alpha = 0.05
)
}
\arguments{
\item{posterior_list}{posterior samples  (class "\code{dlimIM}")}

\item{chain}{if \code{posterior_list} has multiple chains, indicates which chain (class "\code{numeric}")}

\item{burnin}{number of MCMC samples to remove as warm-up (class "\code{numeric}")}

\item{thin}{number post-burn-in MCMC samples to thin by (class "\code{numeric}")}

\item{new_mods}{matrix of new modifier values for prediction ((class "\code{matrix}"))}

\item{m_star}{vector of new modifier index values for prediction (class "\code{numeric}")}

\item{alpha}{significance level (class "\code{numeric}")}
}
\value{
This function returns a list of predicted values and credible bounds \code{list}
\item{betas_cumul}{cumulative effect estimates for each modifier index value (class "\code{numeric}")}
\item{betas_LB}{lower bound for cumulative effect estimates for each modifier index value (class "\code{numeric}")}
\item{betas_UB}{upper bound for cumulative effect estimates for each modifier index value (class "\code{numeric}")}
\item{betas}{point-wise effect estimates for each modifier index value and exposure-time point (class "\code{numeric}")}
\item{betas}{lower bound for point-wise effect estimates for each modifier index value and exposure-time point (class "\code{numeric}")}
\item{betas}{upper bound for point-wise effect estimates for each modifier index value and exposure-time point (class "\code{numeric}")}
}
\description{
Predicted values based on a \code{dlimIM} object.
}
\seealso{
\link[dlimIM]{plot_bdlim}

\link[dlimIM]{MCMC_sampler_m}
}
